Add Routing for IDEAS and ADD IDEA
