+++ src/app/App.js
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from 'react';
-// import { BrowserRouter as Router, Route } from 'react-router-dom'
+import { BrowserRouter as Router, Route } from 'react-router-dom';
 import styled from 'styled-components';
 import uid from 'uid';
 import { getLocal, setLocal } from './services';
 import GlobalStyles from '../misc/GlobalStyles';
 import Header from './Header';
 import Footer from './Footer';
-import IdeasFeed from './IdeasFeed';
-import IdeaForm from './IdeaForm';
+import IdeasFeed from './main/ideas/IdeasFeed';
+import IdeaForm from './main/ideas/CreateIdea';
 
 const Grid = styled.div`
   display: grid;
@@ -43,19 +43,44 @@ export default function App() {
 
   useEffect(() => setLocal('ideas', ideas), [ideas]);
 
-  function onSubmitIdea(newIdea) {
+  function splitToArray(tagString) {
+    return tagString.split(',').map(tag => tag.trim());
+  }
+
+  function handleSubmit(event, history) {
+    event.preventDefault();
+    const form = event.target;
+    const newIdea = {
+      id: uid(),
+      title: form.title.value,
+      text: form.text.value,
+      tags: splitToArray(form.tags.value),
+    };
     setIdeas([newIdea, ...ideas]);
+    history.push('/ideas');
   }
 
   return (
-    <Grid>
-      <GlobalStyles />
-      <Header />
-      <GridArea>
-        <IdeaForm onSubmitIdea={onSubmitIdea} />
-        <IdeasFeed posts={ideas} />
-      </GridArea>
-      <Footer />
-    </Grid>
+    <Router>
+      <Grid>
+        <GlobalStyles />
+        <Header />
+        <GridArea>
+          <Route
+            exact
+            path="/ideas"
+            render={() => <IdeasFeed posts={ideas} />}
+          />
+          <Route
+            exact
+            path="/ideas/add"
+            render={props => (
+              <IdeaForm onSubmit={handleSubmit} history={props.history} />
+            )}
+          />
+        </GridArea>
+        <Footer />
+      </Grid>
+    </Router>
   );
 }
