+++ src/app/main/ideas/CreateIdea.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import styled from 'styled-components';
-import uid from 'uid';
 
 const StyledLabel = styled.label`
   display: grid;
@@ -35,23 +34,9 @@ const StyledForm = styled.form`
   display: grid;
 `;
 
-export default function IdeaForm({ onSubmitIdea }) {
-  function splitToArray(tagString) {
-    return tagString.split(',').map(tag => tag.trim());
-  }
-
-  function handleSubmit(event) {
-    event.preventDefault();
-    const form = event.target;
-    onSubmitIdea({
-      id: uid(),
-      title: form.title.value,
-      text: form.text.value,
-      tags: splitToArray(form.tags.value),
-    });
-  }
+export default function IdeaForm({ onSubmit, history }) {
   return (
-    <StyledForm onSubmit={handleSubmit}>
+    <StyledForm onSubmit={event => onSubmit(event, history)}>
       <StyledLabel>
         title
         <StyledInput
